renaming screenname to anyint
