+++ anyint/solver.py
