+++ anyint/test/anyint_solver_test.py
@@ -3,7 +3,7 @@ import sys, os, pytest
 sys.path.append(os.getcwd())
 
 from do_grader_lib import PartQuality
-from screenname import solver
+from anyint import solver
 
 input_data = ''
