+++ docker/post.sh
@@ -1,6 +1,6 @@
 # Course id :THkzKaBPEeWm9xJA1YXcWQ
 # Item ids:
-#          - FktbJ
+# antint   - FktbJ
 # knapsack - TvXJx
 # coloring - npnKe
 # tsp      - sUgcy
