--- screenname/_coursera
@@ -1,3 +1,0 @@
-heD35Vv_EeajDAoJAOJJEw
-Screen Name
-mJWtY, _coursera, solver.py, Update Screen Name
